==> notebooks/01_split_data.Rmd <==
---
title: "R Notebook"
output: html_notebook
---
## Load Required Libraries
```{r message=TRUE, warning=FALSE}
library(tidyverse)
library(tidymodels)
set.seed(400)
```

## Read Data
```{r}
# https://www.kaggle.com/datasets/kamilpytlak/personal-key-indicators-of-heart-disease
data_original <- read_csv("../data/heart_2020_cleaned.csv")
```

```{r}
data_original <- data_original %>%
  mutate(HeartDisease = factor(HeartDisease, levels = c("Yes", "No")))
```

```{r}
data_original
```

## Create Exploration set
```{r}
data_split <- initial_split(data_original, prop = 0.95)
data_remaining <- training(data_split)
data_exploration <- testing(data_split)
```

## Check distribution of label (`HeartDisease`)
```{r}
data_exploration %>%
  ggplot() +
    geom_bar(mapping = aes(x = HeartDisease))

ggsave(filename = "HeartDisease.png", path = "../results/exploration_plots", 
       device = "png")
```

## Create Train and Test set
```{r}
data_split <- initial_split(data_remaining, prop = 0.95, strata = HeartDisease)
data_train <- training(data_split)
data_test <- testing(data_split)
```

## Summarize Size of Splits
```{r}
print(paste("Full Set Size       :", nrow(data_original)))
print(paste("Exploration Set Size:", nrow(data_exploration)))
print(paste("Test Set Size:       ", nrow(data_test)))
print(paste("Train Set Size      :", nrow(data_train)))
```

## Save Splits
```{r}
data_exploration %>% write_csv(file = "../data/splits/data_exploration.csv")
data_test %>% write_csv(file = "../data/splits/data_test.csv")
data_train %>% write_csv(file = "../data/splits/data_train.csv")
```
