==> notebooks/02_explore_data.Rmd <==
---
title: "R Notebook"
output: html_notebook
---
## Load Required Libraries
```{r message=TRUE, warning=FALSE}
library(tidyverse)
source("../scripts/visualize_categorical.R")
source("../scripts/visualize_numerical.R")
set.seed(400)
```

## Read Exploration Data
```{r}
data_exploration <- read_csv("../data/splits/data_exploration.csv") %>%
  mutate(HeartDisease = factor(HeartDisease, levels = c("Yes", "No")))
```

## Count Number of Missing Values
```{r}
data_exploration %>%
  sapply(function(x) sum(is.na(x)))
```

## Explore Potential Features

### Visualize Categorical Features     
```{r}
data_exploration %>%
  visualize_categorical(Smoking)

ggsave(filename = "Smoking.png", path = "../results/exploration_plots", 
       device = "png")
```

```{r}
data_exploration %>%
  visualize_categorical(AlcoholDrinking)

ggsave(filename = "AlcoholDrinking.png", path = "../results/exploration_plots", 
       device = "png")
```

```{r}
data_exploration %>%
  visualize_categorical(Stroke)

ggsave(filename = "Stroke.png", path = "../results/exploration_plots", 
       device = "png")
```

```{r}
data_exploration %>%
  visualize_categorical(DiffWalking)

ggsave(filename = "DiffWalking.png", path = "../results/exploration_plots", 
       device = "png")
```

```{r}
data_exploration %>%
  visualize_categorical(Sex)

ggsave(filename = "Sex.png", path = "../results/exploration_plots", device = "png")
```

```{r}
data_exploration %>%
  visualize_categorical(AgeCategory)

ggsave(filename = "AgeCategory.png", path = "../results/exploration_plots", 
       device = "png")
```

```{r}
data_exploration %>%
  visualize_categorical(Race)

ggsave(filename = "Race.png", path = "../results/exploration_plots", 
       device = "png")
```

```{r}
data_exploration %>%
  visualize_categorical(Diabetic)

ggsave(filename = "Diabetic.png", path = "../results/exploration_plots", 
       device = "png")
```

```{r}
data_exploration %>%
  visualize_categorical(PhysicalActivity)

ggsave(filename = "PhysicalActivity.png", path = "../results/exploration_plots", 
       device = "png")
```

```{r}
data_exploration %>%
  visualize_categorical(GenHealth)

ggsave(filename = "GenHealth.png", path = "../results/exploration_plots", 
       device = "png")
```

```{r}
data_exploration %>%
  visualize_categorical(Asthma)

ggsave(filename = "Asthma.png", path = "../results/exploration_plots", 
       device = "png")
```

```{r}
data_exploration %>%
  visualize_categorical(KidneyDisease)

ggsave(filename = "KidneyDisease.png", path = "../results/exploration_plots", 
       device = "png")
```

```{r}
data_exploration %>%
  visualize_categorical(SkinCancer)

ggsave(filename = "SkinCancer.png", path = "../results/exploration_plots", 
       device = "png")
```

### Visualize Numerical Features
```{r}
data_exploration %>%
  visualize_numerical(BMI)

ggsave(filename = "BMI.png", path = "../results/exploration_plots", device = "png")
```

```{r}
data_exploration %>%
  visualize_numerical(PhysicalHealth)

ggsave(filename = "PhysicalHealth.png", path = "../results/exploration_plots", 
       device = "png")
```

```{r}
data_exploration %>%
  visualize_numerical(MentalHealth)

ggsave(filename = "MentalHealth.png", path = "../results/exploration_plots", device = "png")
```

```{r}
data_exploration %>%
  visualize_numerical(SleepTime)

ggsave(filename = "SleepTime.png", path = "../results/exploration_plots", device = "png")
```
